Reduce: Add support for USM host memory
The overheads of managing the result memory are significant for the
reduce pattern. USM host memory seems beneficial on Intel GPUs. This
patch uses USM host memory if the synchronous algorithm is used, the
device is an Intel GPU, and the device supports USM host memory. All
other code paths use a result buffer.